PCP: write htoprc DynamicColumns fields using symbolic names
Wework parts of the new DynamicColumns code to read & write
the settings file (htoprc) not using the numeric identifier
for dynamic columns.  Regular static columns are unchanged.

While there, change the PCP columns file parser so that its
syntax is simpler - each column can have one associated PCP
metric only.  The assignment of column identifiers is a bit
simplified also.

Code is added to only enable metric sampling for the metric
behind a dynamic column once that column has been selected.
In addition the way in which values for DynamicColumns were
associated with each process is reworked to resolve compiler
warnings and so we don't keep multiple copies.